+++ packages/sdks/src/functions/is-editing.ts
@@ -1,10 +1,13 @@
 import { TARGET } from '../constants/target.js';
+import { getSearchString, type Search } from '../helpers/search/search.js';
 import { isIframe } from './is-iframe.js';
 
-export function isEditing(): boolean {
+export function isEditing(search?: Search) {
   return (
     isIframe() &&
     (TARGET === 'reactNative' ||
-      window.location.search.indexOf('builder.frameEditing=') !== -1)
+      getSearchString(search || window.location.search).indexOf(
+        'builder.frameEditing='
+      ) !== -1)
   );
 }
