+++ packages/sdks/src/functions/is-previewing.ts
@@ -1,14 +1,18 @@
+import type { Search } from '../helpers/search/search.js';
+import { getSearchString } from '../helpers/search/search.js';
 import { isBrowser } from './is-browser.js';
 import { isEditing } from './is-editing.js';
 
-export function isPreviewing() {
+export function isPreviewing(search?: Search) {
   if (!isBrowser()) {
     return false;
   }
 
-  if (isEditing()) {
+  const normalizedSearch = getSearchString(search || window.location.search);
+
+  if (isEditing(normalizedSearch)) {
     return false;
   }
 
-  return Boolean(location.search.indexOf('builder.preview=') !== -1);
+  return Boolean(normalizedSearch.indexOf('builder.preview=') !== -1);
 }
